+++ azext_edge/edge/providers/check/base.py
@@ -18,6 +18,8 @@ from kubernetes.client.models import (
 from rich.console import Console, NewLine
 from rich.padding import Padding
 
+from .check_manager import CheckManager
+from .base_nodes import check_nodes
 from .common import ALL_NAMESPACES_TARGET, PADDING_SIZE, CoreServiceResourceKinds, ResourceOutputDetailLevel
 from ...common import CheckTaskStatus, ListableEnum
 
@@ -308,82 +310,6 @@ def check_k8s_version(as_list: bool = False) -> Dict[str, Any]:
     return check_manager.as_dict(as_list)
 
 
-def check_nodes(as_list: bool = False) -> Dict[str, Any]:
-    from kubernetes.client.models import V1Node, V1NodeList
-
-    check_manager = CheckManager(check_name="evalClusterNodes", check_desc="Evaluate cluster nodes")
-    target_minimum_nodes = "cluster/nodes"
-    check_manager.add_target(
-        target_name=target_minimum_nodes,
-        conditions=[
-            "len(cluster/nodes)>=1",
-            "(cluster/nodes).each(node.status.allocatable[memory]>=140MiB)",
-        ],
-    )
-
-    try:
-        core_client = client.CoreV1Api()
-        nodes: V1NodeList = core_client.list_node()
-    except ApiException as ae:
-        logger.debug(str(ae))
-        api_error_text = "Unable to fetch nodes. Is there connectivity to the cluster?"
-        check_manager.add_target_eval(
-            target_name=target_minimum_nodes,
-            status=CheckTaskStatus.error.value,
-            value=api_error_text,
-        )
-        check_manager.add_display(
-            target_name=target_minimum_nodes,
-            display=Padding(api_error_text, (0, 0, 0, 8)),
-        )
-    else:
-        node_items: List[V1Node] = nodes.items
-        node_count = len(node_items)
-        target_display = "At least 1 node is required. {}"
-        if node_count < 1:
-            target_display = Padding(
-                target_display.format(f"[red]Detected {node_count}[/red]."),
-                (0, 0, 0, 8),
-            )
-            check_manager.add_target_eval(target_name=target_minimum_nodes, status=CheckTaskStatus.error.value)
-            check_manager.add_display(target_name=target_minimum_nodes, display=target_display)
-            return check_manager.as_dict()
-
-        target_display = Padding(
-            target_display.format(f"[green]Detected {node_count}[/green]."),
-            (0, 0, 0, 8),
-        )
-        check_manager.add_display(target_name=target_minimum_nodes, display=target_display)
-        check_manager.add_display(target_name=target_minimum_nodes, display=NewLine())
-
-        for node in node_items:
-            node_memory_value = {}
-            memory_status = CheckTaskStatus.success.value
-            memory: str = node.status.allocatable["memory"]
-            memory = memory.replace("Ki", "")
-            memory: int = int(int(memory) / 1024)
-            mem_colored = f"[green]{memory}[/green]"
-            node_name = node.metadata.name
-            node_memory_value[node_name] = f"{memory}MiB"
-
-            if memory < 140:
-                memory_status = CheckTaskStatus.warning.value
-                mem_colored = f"[yellow]{memory}[/yellow]"
-
-            node_memory_display = Padding(
-                f"[bright_blue]{node_name}[/bright_blue] {mem_colored} MiB",
-                (0, 0, 0, 8),
-            )
-            check_manager.add_target_eval(
-                target_name=target_minimum_nodes,
-                status=memory_status,
-                value=node_memory_value,
-            )
-            check_manager.add_display(target_name=target_minimum_nodes, display=node_memory_display)
-
-    return check_manager.as_dict(as_list)
-
-
 def decorate_resource_status(status: str) -> str:
     from ...common import ResourceState
 
@@ -407,144 +333,6 @@ def decorate_pod_phase(phase: str) -> Tuple[str, str]:
     if not phase or phase in [PodState.unknown.value, PodState.pending.value]:
         return f"[yellow]{phase}[/yellow]", CheckTaskStatus.warning.value
     return f"[green]{phase}[/green]", CheckTaskStatus.success.value
-
-
-class CheckManager:
-    """
-    {
-        "name":"evaluateBrokerListeners",
-        "description": "Evaluate MQ broker listeners",
-        "status": "warning",
-        "targets": {
-            "mq.iotoperations.azure.com/v1beta1": {
-                "_all_": {
-                    "conditions": null,
-                    "evaluations": [
-                        {
-                            "status": "success"
-                            ...
-                        }
-                    ],
-                }
-            },
-            "brokerlisteners.mq.iotoperations.azure.com": {
-                "default": {
-                    "displays": [],
-                    "conditions": [
-                        "len(brokerlisteners)>=1",
-                        "spec",
-                        "valid(spec.brokerRef)"
-                        ...
-                    ],
-                    "evaluations": [
-                        {
-                            "name": "listener",
-                            "kind": "brokerListener",
-                            "value": {
-                                "spec": { ... },
-                                "valid(spec.brokerRef)": true
-                            },
-                            "status": "warning"
-                        }
-                    ],
-                    "status": "warning"
-                },
-                "other-namespace": {
-                    "displays": [],
-                    "conditions": []
-                    ...
-                }
-            }
-        }
-    }
-    """
-
-    def __init__(self, check_name: str, check_desc: str):
-        self.check_name = check_name
-        self.check_desc = check_desc
-        self.targets = {}
-        self.target_displays = {}
-        self.worst_status = CheckTaskStatus.success.value
-
-    def add_target(self, target_name: str, namespace: str = ALL_NAMESPACES_TARGET, conditions: List[str] = None, description: str = None) -> None:
-        if target_name not in self.targets:
-            # Create a default `None` namespace target for targets with no namespace
-            self.targets[target_name] = {}
-        if namespace and namespace not in self.targets[target_name]:
-            self.targets[target_name][namespace] = {}
-        self.targets[target_name][namespace]["conditions"] = conditions
-        self.targets[target_name][namespace]["evaluations"]: List[dict] = []
-        self.targets[target_name][namespace]["status"] = CheckTaskStatus.success.value
-        if description:
-            self.targets[target_name][namespace]["description"] = description
-
-    def set_target_conditions(self, target_name: str, conditions: List[str], namespace: str = ALL_NAMESPACES_TARGET) -> None:
-        self.targets[target_name][namespace]["conditions"] = conditions
-
-    def add_target_conditions(self, target_name: str, conditions: List[str], namespace: str = ALL_NAMESPACES_TARGET) -> None:
-        self.targets[target_name][namespace]["conditions"].extend(conditions)
-
-    def set_target_status(self, target_name: str, status: str, namespace: str = ALL_NAMESPACES_TARGET) -> None:
-        self._process_status(target_name=target_name, namespace=namespace, status=status)
-
-    def add_target_eval(
-        self,
-        target_name: str,
-        status: str,
-        value: Optional[Any] = None,
-        namespace: str = ALL_NAMESPACES_TARGET,
-        resource_name: Optional[str] = None,
-        resource_kind: Optional[str] = None,
-    ) -> None:
-        eval_dict = {"status": status}
-        if resource_name:
-            eval_dict["name"] = resource_name
-        if value:
-            eval_dict["value"] = value
-        if resource_kind:
-            eval_dict["kind"] = resource_kind
-        self.targets[target_name][namespace]["evaluations"].append(eval_dict)
-        self._process_status(target_name, status, namespace)
-
-    def _process_status(self, target_name: str, status: str, namespace: str = ALL_NAMESPACES_TARGET) -> None:
-        existing_status = self.targets[target_name].get("status", CheckTaskStatus.success.value)
-        if existing_status != status:
-            if existing_status == CheckTaskStatus.success.value and status in [
-                CheckTaskStatus.warning.value,
-                CheckTaskStatus.error.value,
-                CheckTaskStatus.skipped.value,
-            ]:
-                self.targets[target_name][namespace]["status"] = status
-                self.worst_status = status
-            elif (
-                existing_status == CheckTaskStatus.warning.value or existing_status == CheckTaskStatus.skipped.value
-            ) and status in [CheckTaskStatus.error.value]:
-                self.targets[target_name][namespace]["status"] = status
-                self.worst_status = status
-
-    def add_display(self, target_name: str, display: Any, namespace: str = ALL_NAMESPACES_TARGET) -> None:
-        if target_name not in self.target_displays:
-            self.target_displays[target_name] = {}
-        if namespace not in self.target_displays[target_name]:
-            self.target_displays[target_name][namespace] = []
-        self.target_displays[target_name][namespace].append(display)
-
-    def as_dict(self, as_list: bool = False) -> Dict[str, Any]:
-        from copy import deepcopy
-
-        result = {
-            "name": self.check_name,
-            "description": self.check_desc,
-            "targets": {},
-            "status": self.worst_status,
-        }
-        result["targets"] = deepcopy(self.targets)
-        if as_list:
-            for type in self.target_displays:
-                for namespace in self.target_displays[type]:
-                    result["targets"][type][namespace]["displays"] = deepcopy(self.target_displays[type][namespace])
-
-        return result
 
 
 def evaluate_pod_health(
