+++ azext_edge/edge/providers/checks.py
@@ -43,6 +43,8 @@ def run_checks(
     if resource_kinds:
         _validate_resource_kinds_under_service(ops_service, resource_kinds)
 
+    check_pre_deployment(result, as_list)
+    return result
     with console.status(status="Analyzing cluster...", refresh_per_second=12.5):
         from time import sleep
 
@@ -71,7 +73,7 @@ def run_checks(
             )
 
         if as_list:
-            return process_as_list(console=console, result=result) if as_list else result
+            return process_as_list(console=console, result=result)
         return result
